+++ tests/bas_metadata_library/test_standard_iso_19115_1.py
@@ -24,7 +24,6 @@ from lxml.etree import ElementTree, XML, fromstring, tostring, XMLParser
 from bas_metadata_library import RecordValidationError
 from bas_metadata_library.standards.iso_19115_1 import (
     Namespaces,
-    MetadataRecordConfigV1,
     MetadataRecordConfigV2,
     MetadataRecord,
 )
@@ -38,7 +37,6 @@ from tests.bas_metadata_library.standard_iso_19115_1_common import (
     assert_identifier,
 )
 from tests.resources.configs.iso19115_1_standard import (
-    configs_safe_v1,
     configs_safe_v2,
     configs_v2_all,
 )
@@ -50,44 +48,12 @@ standard = "iso-19115-1"
 namespaces = Namespaces()
 
 
-def test_invalid_configuration_v1():
-    config = {"invalid-configuration": "invalid-configuration"}
-    with pytest.raises(ValidationError) as e:
-        configuration = MetadataRecordConfigV1(**config)
-        configuration.validate()
-    assert "'language' is a required property" in str(e.value)
-
-
 def test_invalid_configuration_v2():
     config = {"invalid-configuration": "invalid-configuration"}
     with pytest.raises(ValidationError) as e:
         configuration = MetadataRecordConfigV2(**config)
         configuration.validate()
     assert "'hierarchy_level' is a required property" in str(e.value)
-
-
-def test_configuration_v1_from_json_file():
-    configuration = MetadataRecordConfigV1()
-    configuration.load(file=Path("tests/resources/configs/iso19115_1_standard_minimal_record_v1.json"))
-    configuration.validate()
-    _config = deepcopy(configs_safe_v1["minimal_v1"])
-    _config["resource"]["dates"].append(
-        {"date": datetime.datetime(2018, 1, 1, 10, 0, 0, tzinfo=datetime.timezone.utc), "date_type": "revision"}
-    )
-    assert configuration.config == _config
-
-
-def test_configuration_v1_from_json_string():
-    with open(str(Path("tests/resources/configs/iso19115_1_standard_minimal_record_v1.json")), mode="r") as file:
-        config_str = file.read()
-        configuration = MetadataRecordConfigV1()
-        configuration.loads(string=config_str)
-        configuration.validate()
-        _config = deepcopy(configs_safe_v1["minimal_v1"])
-        _config["resource"]["dates"].append(
-            {"date": datetime.datetime(2018, 1, 1, 10, 0, 0, tzinfo=datetime.timezone.utc), "date_type": "revision"}
-        )
-        assert configuration.config == _config
 
 
 def test_configuration_v2_from_json_file():
@@ -1560,9 +1526,7 @@ def test_edge_case_distribution_option_transfer_options_no_properties():
     )
     _record = MetadataRecord(record=record)
     _config = _record.make_config()
-    assert _record.make_config().config["distribution"][0]["distribution_options"] == [
-        {"transfer_option": {"format": "netCDF"}}
-    ]
+    assert _record.make_config().config["distribution"][0]["distribution_options"] == [{"format": {"format": "netCDF"}}]
 
 
 def test_edge_case_distribution_option_no_id():
@@ -1642,17 +1606,13 @@ def test_edge_case_distribution_option_more_formats_than_transfer_options():
             },
             "distribution_options": [
                 {
-                    "format": {"format": "netCDF"},
-                    "transfer_option": {
-                        "online_resource": {
-                            "href": "https://ramadda.data.bas.ac.uk/repository/entry/show?entryid=b1a7d1b5-c419-41e7-9178-b1ffd76d5371",
-                            "title": "Get Data",
-                            "description": "Download measurement data",
-                            "function": "download",
-                        }
-                    },
+                    "format": {"format": "blue"},
+                    "transfer_option": {"online_resource": {"href": "https://example.com/blue"}},
                 },
-                {"format": {"format": "netCDF-x"}, "transfer_option": {"online_resource": {"href": ""}}},
+                {
+                    "format": {"format": "red"},
+                    "transfer_option": {"online_resource": {"href": "https://example.com/red"}},
+                },
             ],
         }
     ]
@@ -1664,17 +1624,56 @@ def test_edge_case_distribution_option_more_formats_than_transfer_options():
     _config = _record.make_config()
     assert _config.config["distribution"][0]["distribution_options"] == [
         {
-            "format": {"format": "netCDF"},
-            "transfer_option": {
-                "online_resource": {
-                    "href": "https://ramadda.data.bas.ac.uk/repository/entry/show?entryid=b1a7d1b5-c419-41e7-9178-b1ffd76d5371",
-                    "title": "Get Data",
-                    "description": "Download measurement data",
-                    "function": "download",
-                }
-            },
+            "format": {"format": "blue"},
+            "transfer_option": {"online_resource": {"href": "https://example.com/blue"}},
         },
-        {"transfer_option": {"format": "netCDF-x"}},
+        {
+            "format": {"format": "red"},
+        },
+    ]
+
+
+def test_edge_case_distribution_option_more_transfer_options_than_formats():
+    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config["distribution"] = [
+        {
+            "distributor": {
+                "organisation": {"name": "UK Polar Data Centre"},
+                "phone": "+44 (0)1223 221400",
+                "address": {
+                    "delivery_point": "British Antarctic Survey, High Cross, Madingley Road",
+                    "city": "Cambridge",
+                    "administrative_area": "Cambridgeshire",
+                    "postal_code": "CB3 0ET",
+                    "country": "United Kingdom",
+                },
+                "email": "polardatacentre@bas.ac.uk",
+                "role": ["distributor"],
+            },
+            "distribution_options": [
+                {
+                    "format": {"format": "blue"},
+                    "transfer_option": {"online_resource": {"href": "https://example.com/blue"}},
+                },
+                {
+                    "format": {"format": "red"},
+                    "transfer_option": {"online_resource": {"href": "https://example.com/red"}},
+                },
+            ],
+        }
+    ]
+    config = MetadataRecordConfigV2(**config)
+    record = MetadataRecord(configuration=config)
+    del record.attributes["distribution"][0]["distribution_options"][1]["format"]
+    record = record.generate_xml_document().decode()
+    _record = MetadataRecord(record=record)
+    _config = _record.make_config()
+    assert _config.config["distribution"][0]["distribution_options"] == [
+        {
+            "format": {"format": "blue"},
+            "transfer_option": {"online_resource": {"href": "https://example.com/blue"}},
+        },
+        {"transfer_option": {"online_resource": {"href": "https://example.com/red"}}},
     ]
 
 
@@ -1833,18 +1832,6 @@ class MockResponse:
         )
 
 
-@pytest.mark.parametrize("config_name", list(configs_safe_v1.keys()))
-def test_parse_existing_record_v1(config_name):
-    with open(f"tests/resources/records/iso-19115-1/{config_name}-record.xml") as record_file:
-        record_data = record_file.read()
-
-    record = MetadataRecord(record=record_data)
-    configuration = MetadataRecordConfigV1()
-    configuration.convert_from_v2_configuration(record.make_config())
-    config = configuration.config
-    assert config == configs_safe_v1[config_name]
-
-
 @pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
 def test_parse_existing_record_v2(config_name):
     with open(f"tests/resources/records/iso-19115-1/{config_name}-record.xml") as record_file:
@@ -1854,29 +1841,6 @@ def test_parse_existing_record_v2(config_name):
     configuration = record.make_config()
     config = configuration.config
     assert config == configs_safe_v2[config_name]
-
-
-@pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v1.keys()))
-def test_lossless_conversion_v1(get_record_response, config_name):
-    _record = tostring(
-        get_record_response(standard=standard, config=config_name),
-        pretty_print=True,
-        xml_declaration=True,
-        encoding="utf-8",
-    ).decode()
-    _config = configs_safe_v1[config_name]
-
-    record = MetadataRecord(record=_record)
-    config_ = MetadataRecordConfigV1()
-    config_.convert_from_v2_configuration(record.make_config())
-    config_ = config_.config
-
-    config = MetadataRecordConfigV1(**config_)
-    config = config.convert_to_v2_configuration()
-    record_ = MetadataRecord(configuration=config).generate_xml_document().decode()
-    assert _record == record_
-    assert _config == config_
 
 
 @pytest.mark.usefixtures("get_record_response")